requirements.txt
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.30.0
pandas>=1.5.0
pyyaml>=6.0
